backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
supabase==2.4.6
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
joblib==1.4.2
shap==0.45.1
requests==2.31.0
apscheduler==3.10.4
google-generativeai==0.5.4
pydantic==2.7.1